🧹 Extract duplicate `setRealmField` logic to `RealmUtils`
