+++ app/src/main/java/org/ole/planet/myplanet/repository/UploadRepositoryImpl.kt
@@ -55,8 +55,8 @@ class UploadRepositoryImpl @Inject constructor(
                     val item = itemsById[uploadedItem.localId]
 
                     item?.let {
-                        setRealmField(it, "_id", uploadedItem.remoteId)
-                        setRealmField(it, "_rev", uploadedItem.remoteRev)
+                        org.ole.planet.myplanet.utils.RealmUtils.setRealmField(it, "_id", uploadedItem.remoteId)
+                        org.ole.planet.myplanet.utils.RealmUtils.setRealmField(it, "_rev", uploadedItem.remoteRev)
                         config.additionalUpdates?.invoke(realm, it, uploadedItem)
                     } ?: run {
                         failedLocally.add(uploadedItem)
@@ -80,44 +80,5 @@ class UploadRepositoryImpl @Inject constructor(
 
     override suspend fun fetchExistingDoc(url: String): Response<JsonObject> {
         return apiInterface.getJsonObject(UrlUtils.header, url)
-    }
-
-    private class FieldCacheEntry(val field: Field?)
-
-    private val fieldCache = ConcurrentHashMap<Pair<Class<*>, String>, FieldCacheEntry>()
-
-    private fun setRealmField(obj: RealmObject, fieldName: String, value: Any?) {
-        try {
-            val cacheKey = Pair(obj.javaClass, fieldName)
-            var entry = fieldCache[cacheKey]
-
-            if (entry == null) {
-                var clazz: Class<*>? = obj.javaClass
-                var field: Field? = null
-
-                while (clazz != null && field == null) {
-                    try {
-                        field = clazz.getDeclaredField(fieldName)
-                    } catch (e: NoSuchFieldException) {
-                        clazz = clazz.superclass
-                    }
-                }
-
-                if (field != null) {
-                    field.isAccessible = true
-                } else {
-                    Log.w("UploadRepositoryImpl", "Field $fieldName not found in class hierarchy of ${obj.javaClass.simpleName}")
-                }
-
-                entry = FieldCacheEntry(field)
-                fieldCache[cacheKey] = entry
-            } else if (entry.field == null) {
-                Log.w("UploadRepositoryImpl", "Field $fieldName not found in class hierarchy of ${obj.javaClass.simpleName}")
-            }
-
-            entry.field?.set(obj, value)
-        } catch (e: Exception) {
-            Log.w("UploadRepositoryImpl", "Failed to set field $fieldName: ${e.message}")
-        }
     }
 }
